- ItemsIncomeDao.kt on insert has to insert another entities: Item, Container, Supplier. Maybe use id's instead of objects in Exposed Entities?

+++ src/main/kotlin/ui/screens/entity_screen_with_filter/entities_list/EntitiesListComponent.kt
@@ -85,13 +85,13 @@ class EntitiesListComponent<T : IEntity<*>>(
         }
     }
 
-    override fun setCurrentPage(currentPage: Int) {
+    override fun setCurrentPage(currentPage: Long) {
         _state.reduce {
             it.copy(pagingParameters = it.pagingParameters.copy(currentPage = currentPage))
         }
     }
 
-    override fun setItemsPerPage(itemsPerPage: Int) {
+    override fun setItemsPerPage(itemsPerPage: Long) {
         _state.reduce {
             it.copy(pagingParameters = it.pagingParameters.copy(itemsPerPage = itemsPerPage))
         }
